Economic/maps.me/requirements/requirements.txt
elasticsearch
requests
kafka-python
